Added check for existing Ubuntu base image

+++ build-locally.sh
@@ -3,11 +3,21 @@
 
 PURPLEBOLD="$(tput setf 5 bold)"
 
+#Checks if image ubuntu already exists:
+
+IMAGEEXISTS=true
+if [[ "$(docker image inspect ubuntu >/dev/null 2>&1 && echo true || echo false)" = "false" ]]; then
+  IMAGEEXISTS=false
+fi
+
 printf "${PURPLEBOLD}Building temporary modified Ubuntu image:\n"
 docker build -t lancachenet/ubuntu:latest --progress tty https://github.com/lancachenet/ubuntu.git
 
-printf "${PURPLEBOLD}Removing standart Ubuntu image:\n"
-docker rmi ubuntu
+#Removes standart Ubuntu image if not present before running:
+if [$IMAGEEXISTS" == false]; then
+  printf "${PURPLEBOLD}Removing standart Ubuntu image:\n"
+  docker rmi ubuntu
+fi
 
 printf  "${PURPLEBOLD}Building temporary Ubuntu-Nginx image:\n"
 docker build -t lancachenet/ubuntu-nginx:latest --progress tty https://github.com/lancachenet/ubuntu-nginx.git
